allow signing non-root nodes
sign! is now a method on Node, instead of Document. this allows you
to place the signature at any level in the document. note that for
partial signing, you need to use a DTD, or call `set_id_attribute`
on the element you're signing to identify the ID, and then pass
the value of that ID as the uri param to sign!.

also combined the two signing methods into one, since they were
mostly identical, and just altered in how they added the key
info.